// extensions/bangumi/src/types/generated.ts
export interface components {
  schemas: {
// ...
    Episode: {
      /** ID */
      id: number
      /**
       * Type
       * @description `0` 本篇，`1` SP，`2` OP，`3` ED
       */
      type: number
      /** Name */
      name: string
      /** Name Cn */
      name_cn: string
      /**
       * Sort
       * @description 同类条目的排序和集数
       */
      sort: number
      /**
       * Ep
       * @description 条目内的集数, 从`1`开始。非本篇剧集的此字段无意义
       */
      ep?: number
      /** Airdate */
      airdate: string
      /** Comment */
      comment: number
      /**
       * Duration
       * @description 维基人填写的原始时长
       */
      duration: string
      /**
       * Desc
       * @description 简介
       */
      desc: string
      /**
       * Disc
       * @description 音乐曲目的碟片数
       */
      disc: number
      /** @description 服务器解析的时长，无法解析时为 `0` */
      duration_seconds?: number
    }
// ...
  }
// ...
}